refactor(gate): split ratchet owners
