+++ crates/tokmd-gate/src/ratchet.rs
@@ -1,221 +1,16 @@
 //! Ratchet rule evaluation logic.
 
 mod change;
+mod evaluate;
+mod policy;
 
-use change::percentage_change;
-
-use crate::numeric::value_to_f64;
-use crate::pointer::resolve_pointer;
-use crate::types::{RatchetConfig, RatchetGateResult, RatchetResult, RatchetRule, RuleLevel};
-use serde_json::Value;
-
-/// Evaluate a single ratchet rule against baseline and current values.
-///
-/// This is a convenience wrapper around [`evaluate_ratchet_with_options`] with
-/// strict missing value handling (fails if baseline or current values are missing).
-///
-/// # Arguments
-/// * `rule` - The ratchet rule to evaluate
-/// * `baseline` - The baseline JSON receipt to compare against
-/// * `current` - The current JSON receipt to check
-///
-/// # Returns
-/// A `RatchetResult` with pass/fail status and detailed information.
-#[cfg_attr(not(test), allow(dead_code))]
-pub(crate) fn evaluate_ratchet(
-    rule: &RatchetRule,
-    baseline: &Value,
-    current: &Value,
-) -> RatchetResult {
-    evaluate_ratchet_with_options(rule, baseline, current, false, false)
-}
-
-/// Evaluate a ratchet rule with configurable missing value handling.
-///
-/// # Arguments
-/// * `rule` - The ratchet rule to evaluate
-/// * `baseline` - The baseline JSON receipt
-/// * `current` - The current JSON receipt
-/// * `allow_missing_baseline` - Treat missing baseline as pass
-/// * `allow_missing_current` - Treat missing current value as pass
-///
-/// # Returns
-/// A `RatchetResult` with pass/fail status.
-pub fn evaluate_ratchet_with_options(
-    rule: &RatchetRule,
-    baseline: &Value,
-    current: &Value,
-    allow_missing_baseline: bool,
-    allow_missing_current: bool,
-) -> RatchetResult {
-    let baseline_resolved = resolve_pointer(baseline, &rule.pointer);
-    let current_resolved = resolve_pointer(current, &rule.pointer);
-
-    // Handle missing current value
-    let current_value = match current_resolved.and_then(value_to_f64) {
-        Some(v) => v,
-        None => {
-            if allow_missing_current {
-                return RatchetResult {
-                    rule: rule.clone(),
-                    passed: true,
-                    baseline_value: baseline_resolved.and_then(value_to_f64),
-                    current_value: f64::NAN,
-                    change_pct: None,
-                    message: format!(
-                        "Current value not found at pointer '{}' (allowed)",
-                        rule.pointer
-                    ),
-                };
-            } else {
-                return RatchetResult {
-                    rule: rule.clone(),
-                    passed: false,
-                    baseline_value: baseline_resolved.and_then(value_to_f64),
-                    current_value: f64::NAN,
-                    change_pct: None,
-                    message: format!(
-                        "Current value not found or not numeric at pointer '{}'",
-                        rule.pointer
-                    ),
-                };
-            }
-        }
-    };
-
-    // Extract baseline value
-    let baseline_value = baseline_resolved.and_then(value_to_f64);
-
-    let change_pct = percentage_change(baseline_value, current_value);
-
-    // Check max_value constraint (absolute ceiling)
-    if let Some(max_val) = rule.max_value
-        && current_value > max_val
-    {
-        return RatchetResult {
-            rule: rule.clone(),
-            passed: false,
-            baseline_value,
-            current_value,
-            change_pct,
-            message: format!(
-                "Current value {} exceeds maximum allowed value {}",
-                current_value, max_val
-            ),
-        };
-    }
-
-    // Check max_increase_pct constraint (relative to baseline)
-    if let Some(max_inc_pct) = rule.max_increase_pct {
-        match baseline_value {
-            Some(_bv) => {
-                let pct = change_pct.unwrap_or(0.0);
-                if pct > max_inc_pct {
-                    return RatchetResult {
-                        rule: rule.clone(),
-                        passed: false,
-                        baseline_value,
-                        current_value,
-                        change_pct,
-                        message: format!(
-                            "Increase of {:.2}% exceeds maximum allowed increase of {:.2}%",
-                            pct, max_inc_pct
-                        ),
-                    };
-                }
-            }
-            None => {
-                // No baseline value - can't evaluate percentage increase
-                if allow_missing_baseline {
-                    let desc = rule
-                        .description
-                        .as_deref()
-                        .unwrap_or("Ratchet check passed (no baseline)");
-                    return RatchetResult {
-                        rule: rule.clone(),
-                        passed: true,
-                        baseline_value: None,
-                        current_value,
-                        change_pct: None,
-                        message: format!("{}: current value = {}", desc, current_value),
-                    };
-                } else {
-                    return RatchetResult {
-                        rule: rule.clone(),
-                        passed: false,
-                        baseline_value: None,
-                        current_value,
-                        change_pct: None,
-                        message: format!(
-                            "Baseline value not found at pointer '{}', cannot evaluate percentage increase",
-                            rule.pointer
-                        ),
-                    };
-                }
-            }
-        }
-    }
-
-    // All checks passed
-    let desc = rule
-        .description
-        .as_deref()
-        .unwrap_or("Ratchet check passed");
-
-    let message = match (baseline_value, change_pct) {
-        (Some(bv), Some(pct)) => format!("{}: {} -> {} ({:+.2}%)", desc, bv, current_value, pct),
-        _ => format!("{}: current value = {}", desc, current_value),
-    };
-
-    RatchetResult {
-        rule: rule.clone(),
-        passed: true,
-        baseline_value,
-        current_value,
-        change_pct,
-        message,
-    }
-}
-
-/// Evaluate all ratchet rules against baseline and current receipts.
-///
-/// # Arguments
-/// * `config` - The ratchet configuration with rules
-/// * `baseline` - The baseline JSON receipt
-/// * `current` - The current JSON receipt
-///
-/// # Returns
-/// A `RatchetGateResult` with overall pass/fail and individual results.
-pub fn evaluate_ratchet_policy(
-    config: &RatchetConfig,
-    baseline: &Value,
-    current: &Value,
-) -> RatchetGateResult {
-    let mut ratchet_results = Vec::with_capacity(config.rules.len());
-
-    for rule in &config.rules {
-        let result = evaluate_ratchet_with_options(
-            rule,
-            baseline,
-            current,
-            config.allow_missing_baseline,
-            config.allow_missing_current,
-        );
-
-        let failed_error = !result.passed && result.rule.level == RuleLevel::Error;
-        ratchet_results.push(result);
-
-        if config.fail_fast && failed_error {
-            break;
-        }
-    }
-
-    RatchetGateResult::from_results(ratchet_results)
-}
+pub use policy::evaluate_ratchet_policy;
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::types::{RatchetConfig, RatchetRule, RuleLevel};
+    use evaluate::{evaluate_ratchet, evaluate_ratchet_with_options};
     use serde_json::json;
 
     fn make_rule(
